Search

+++ Crawler/HorsePostgresDB.py
@@ -1,12 +1,12 @@
 """
-Abstractions over the use of the HorseCrawler database
+Abstractions over the use of the Crawler database
 """
 
 from urllib.parse import urlparse
 import json
 import time
-import mysql
-import mysql.connector
+import psycopg2
+import psycopg2.extras
 
 
 class HorseDB:
@@ -22,29 +22,30 @@ class HorseDB:
 
     @staticmethod
     def getDatabaseConn():
-        return mysql.connector.connect(database='db', user='root', passwd='root', host='localhost', port=3306)
-        return mysql.connector.connect(database='db', user='postgres', passwd='root', host='antonchristensen.net', port=3303)
+        return psycopg2.connect(dbname='db', user='user', password='pass', host='localhost')
+        return psycopg2.connect(dbname='db', user='postgres', password='root', host='antonchristensen.net')
 
     def tableExists(self, tableName):
         c = self.dbconn.cursor()
-        c.execute("SELECT * FROM information_schema.tables WHERE table_name = '{}'".format(tableName))
+        c.execute("SELECT * FROM information_schema.tables WHERE table_name=%s;", (tableName,))
 
         return len(c.fetchall()) > 0
 
     def dropTable(self, tableName):
         c = self.dbconn.cursor()
-        c.execute("DROP TABLE IF EXISTS {}".format(tableName))
+        c.execute("DROP TABLE " + tableName + ";")
         self.dbconn.commit()
 
     def rebuildDatabase(self):
-        tables = ['pages', 'hosts', 'q', 'links', 'reverse_index']
+        tables = ['pages', 'hosts', 'q', 'links', 'index']
         for t in tables:
-            self.dropTable(t)
+            if self.tableExists(t):
+                self.dropTable(t)
 
         self.createDatabase()
 
     def databaseExists(self):
-        tables = ['pages', 'hosts', 'q', 'links', 'reverse_index']
+        tables = ['pages', 'hosts', 'q', 'links', 'index']
         for t in tables:
             if not self.tableExists(t):
                 return False
@@ -57,45 +58,45 @@ class HorseDB:
         c = self.dbconn.cursor()
         c.execute(''' 
             CREATE TABLE pages (
-                id            SERIAL        NOT NULL PRIMARY KEY  UNIQUE,
-                url           VARCHAR(1024) NOT NULL,
-                lang          CHAR(2)       NOT NULL,
-                last_visited  DOUBLE        NOT NULL
+                id            SERIAL    NOT NULL PRIMARY KEY  UNIQUE,
+                url           VARCHAR   NOT NULL,
+                lang          VARCHAR(2) NOT NULL,
+                last_visited  NUMERIC   NOT NULL
             );
         ''')
 
         c.execute(''' 
             CREATE TABLE hosts (
-                id                      SERIAL          NOT NULL PRIMARY KEY UNIQUE,
-                host                    VARCHAR(255)    NOT NULL UNIQUE,
-                last_visited            DOUBLE          NOT NULL,
-                disallow_list           TEXT            NOT NULL,
-                disallow_list_updated   DOUBLE          NOT NULL
+                id                      SERIAL  NOT NULL PRIMARY KEY UNIQUE,
+                host                    VARCHAR NOT NULL UNIQUE,
+                last_visited            NUMERIC NOT NULL,
+                disallow_list           VARCHAR NOT NULL,
+                disallow_list_updated   NUMERIC NOT NULL
             );
         ''')
 
         c.execute(''' 
             CREATE TABLE q (
-                id        SERIAL        NOT NULL PRIMARY KEY UNIQUE,
-                url       VARCHAR(1024) NOT NULL,
-                host      VARCHAR(255)  NOT NULL
+                id        SERIAL    NOT NULL PRIMARY KEY UNIQUE,
+                url       VARCHAR   NOT NULL,
+                host      VARCHAR   NOT NULL
             );
         ''')
 
         c.execute(''' 
             CREATE TABLE links (
-                id            SERIAL        PRIMARY KEY,
-                from_page_id  INT           NOT NULL,
-                to_page_url   VARCHAR(1024) NOT NULL
+                from_page_id  INT       NOT NULL,
+                to_page_url   VARCHAR   NOT NULL,
+                PRIMARY KEY (from_page_id, to_page_url)
             );
         ''')
 
         c.execute(''' 
-            CREATE TABLE reverse_index (
-                id        SERIAL        NOT NULL PRIMARY KEY UNIQUE,
-                term      VARCHAR(100)  NOT NULL,
-                frequency INT           NOT NULL,
-                page_id   INT           NOT NULL
+            CREATE TABLE index (
+                id        SERIAL    NOT NULL PRIMARY KEY UNIQUE,
+                term      VARCHAR   NOT NULL,
+                frequency INT       NOT NULL,
+                page_id   INT       NOT NULL
             );
         ''')
 
@@ -109,7 +110,7 @@ class HorseDB:
         oneWeek = 60 * 60 * 24 * 7
 
         c = self.dbconn.cursor()
-        c.execute("SELECT id, disallow_list, disallow_list_updated FROM hosts WHERE host = '{}'".format(parsedUrl.netloc))
+        c.execute('SELECT id, disallow_list, disallow_list_updated FROM hosts WHERE host = %s', (parsedUrl.netloc,))
         results = c.fetchone()
 
         if results is None:
@@ -126,10 +127,10 @@ class HorseDB:
         c = self.dbconn.cursor()
 
         try:
-            c.execute("INSERT INTO hosts (host, last_visited, disallow_list, disallow_list_updated) VALUES ('{}', {}, '{}', {})".format(parsedUrl.netloc, 0, '', 0))
+            c.execute('INSERT INTO hosts (host, last_visited, disallow_list, disallow_list_updated) VALUES (%s, %s, %s, %s)', (parsedUrl.netloc, 0, '', 0))
 
             self.dbconn.commit()
-        except Exception:
+        except psycopg2.InternalError:
             print('Does this ever happen?!?!?')
             pass
 
@@ -139,7 +140,7 @@ class HorseDB:
         host = urlparse(url).netloc
 
         c = self.dbconn.cursor()
-        c.execute("UPDATE hosts SET disallow_list = '{}', disallow_list_updated = {} WHERE host = '{}'".format(encodedList, now, host))
+        c.execute('UPDATE hosts SET disallow_list = %s, disallow_list_updated = %s WHERE host = %s', (encodedList, now, host,))
         self.dbconn.commit()
 
     def updateHostVisitTime(self, url):
@@ -147,12 +148,12 @@ class HorseDB:
         host = urlparse(url).netloc
 
         c = self.dbconn.cursor()
-        c.execute("UPDATE hosts SET last_visited = {} WHERE host = '{}'".format(now, host))
+        c.execute('UPDATE hosts SET last_visited = %s WHERE host = %s', (now, host,))
         self.dbconn.commit()
 
     def hasUrlBeenCrawledRecently(self, url):
         c = self.dbconn.cursor()
-        c.execute("SELECT last_visited FROM pages WHERE url = '{}'".format(url))
+        c.execute('SELECT last_visited FROM pages WHERE url = %s', (url,))
 
         last_visited = c.fetchone()
 
@@ -176,7 +177,7 @@ class HorseDB:
 
     def getPageId(self, url):
         c = self.dbconn.cursor()
-        c.execute("SELECT * FROM pages WHERE url = '{}'".format(url))
+        c.execute('SELECT * FROM pages WHERE url = %s', (url,))
         results = c.fetchone()
 
         if results is None:
@@ -188,33 +189,33 @@ class HorseDB:
         now = time.time()
 
         c = self.dbconn.cursor()
-        c.execute("INSERT INTO pages (url, lang, last_visited) VALUES (%s, %s, %s);", (url, lang, now))
+        c.execute('INSERT INTO pages (url, lang, last_visited) VALUES (%s, %s, %s) RETURNING id;', (url, lang, now,))
         self.dbconn.commit()
 
-        return c.lastrowid
+        return c.fetchone()[0]
 
     def updatePage(self, page_id, lang):
         now = time.time()
 
         c = self.dbconn.cursor()
-        c.execute("UPDATE pages SET last_visited = {}, lang = '{}' WHERE id = {}".format(now, lang, page_id))
+        c.execute('UPDATE pages SET last_visited = %s, lang = %s WHERE id = %s', (now, lang, page_id))
         self.dbconn.commit()
 
     def deleteLinks(self, from_page_id):
         c = self.dbconn.cursor()
-        c.execute("DELETE FROM links WHERE from_page_id = {}".format(from_page_id))
+        c.execute('DELETE FROM links WHERE from_page_id = %s', (from_page_id,))
         self.dbconn.commit()
 
     def deleteIndex(self, page_id):
         c = self.dbconn.cursor()
-        c.execute("DELETE FROM reverse_index WHERE page_id = {}".format(page_id,))
+        c.execute('DELETE FROM index WHERE page_id = %s', (page_id,))
         self.dbconn.commit()
 
     def insertLinks(self, from_page_id, links):
         params = [(from_page_id, link) for link in links]
 
         c = self.dbconn.cursor()
-        c.executemany("INSERT INTO links (from_page_id, to_page_url) VALUES (%s, %s)", params)
+        psycopg2.extras.execute_values(c, 'INSERT INTO links (from_page_id, to_page_url) VALUES %s', params, page_size=1000)
         self.dbconn.commit()
 
     def insertIndex(self, page_id, tokens):
@@ -226,12 +227,12 @@ class HorseDB:
         params = [(page_id, pair[0], pair[1]) for pair in params]
 
         c = self.dbconn.cursor()
-        c.executemany("INSERT INTO reverse_index (page_id, term, frequency) VALUES (%s, %s, %s)", params)
+        psycopg2.extras.execute_values(c, 'INSERT INTO index (page_id, term, frequency) VALUES %s', params, page_size=1000)
         self.dbconn.commit()
 
     def isInQueue(self, url):
         c = self.dbconn.cursor()
-        c.execute("SELECT COUNT(*) FROM q WHERE url = '{}'".format(url))
+        c.execute('SELECT COUNT(*) FROM q WHERE url = %s', (url,))
         results = c.fetchone()
         return results[0] > 0
 
@@ -239,7 +240,7 @@ class HorseDB:
         host = urlparse(url).netloc
 
         c = self.dbconn.cursor()
-        c.execute("INSERT INTO q (url, host) VALUES ('{}', '{}')".format(url, host))
+        c.execute('INSERT INTO q (url, host) VALUES (%s, %s)', (url, host))
         self.dbconn.commit()
 
     def dequeue(self):
@@ -252,21 +253,21 @@ class HorseDB:
                 SELECT * FROM q 
                 LEFT JOIN hosts AS h 
                     ON q.host = h.host 
-                WHERE h.last_visited < {} 
+                WHERE h.last_visited < %s 
                 ORDER BY q.id 
-                ASC
-                LIMIT 1;'''.format((now - hostRestitutionTimeInSeconds),))
+                ASC NULLS FIRST
+                LIMIT 1;''', ((now - hostRestitutionTimeInSeconds),))
             row = c.fetchone()
 
             if row is None:
                 if self.qSize() > 0:
                     self.debugService.add('QUEUE', 'We visited everything recently... Lets just visit something again and not care about being so fucking polite')
-                    c.execute("SELECT * FROM q LEFT JOIN hosts AS h ON q.host = h.host ORDER BY last_visited ASC LIMIT 1;")
+                    c.execute('SELECT * FROM q LEFT JOIN hosts AS h ON q.host = h.host ORDER BY last_visited ASC NULLS FIRST LIMIT 1;')
                     row = c.fetchone()
                 else:
                     self.debugService.add('DONE', 'The web has been crawled. No more to see here.')
 
-            c.execute("DELETE FROM q WHERE id = {}".format(row[0]))
+            c.execute('DELETE FROM q WHERE id = %s', (row[0],))
             self.dbconn.commit()
 
             if c.rowcount == 1:
@@ -276,5 +277,6 @@ class HorseDB:
 
     def qSize(self):
         c = self.dbconn.cursor()
-        c.execute("SELECT COUNT(*) FROM q;")
+        c.execute('SELECT COUNT(*) FROM q;')
         return c.fetchone()[0]
+
